+++ frontend/src/util/SmartAuth.js
